pyproject.toml
[project]
name = "aixnda"
version = "0.1.0"
description = "aixnda using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.95.0,<1.0.0",
    "PyMuPDF>=1.18.19",  # Correct version of PyMuPDF (fitz)
    "python-docx>=0.8.11",  # For Word document processing
    "fpdf>=1.7.2",
    "streamlit>=1.41.1",
    "Markdown>=3.7"
]

[project.scripts]
aixnda = "aixnda.main:run"
run_crew = "aixnda.main:run"
train = "aixnda.main:train"
replay = "aixnda.main:replay"
test = "aixnda.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
